bloomreachSkusonPlp/requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
